% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKeys.R
\name{createKeys}
\alias{createKeys}
\title{Creates Unique Keys from Music Brainz Database}
\usage{
createKeys()
}
\value{
A .csv file containing the artist keys and artist names
}
\description{
Creates Unique Keys from Music Brainz Database
}
