% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Create Table of Descriptive Statistics}
\usage{
descriptives()
}
\value{
Table 1, descriptive statistics for Indicative Revenue
}
\description{
\code{descriptives()} returns a table of descriptive statistics of IR for superstar musicians,
for the period 2010-2020.
}
