% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatequantiles.R
\name{estimatequantiles}
\alias{estimatequantiles}
\title{Estimates Quantiles of the distribution by year}
\usage{
estimatequantiles(yr)
}
\arguments{
\item{yr}{The year to estimate the quantiles}

\item{workingdir}{The working directory}
}
\value{
Quantiles: 0.01, 0.1, 0.5, 0.9
}
\description{
Estimates Quantiles of the distribution by year
}
