% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammys.R
\name{grammys}
\alias{grammys}
\title{Grammy Wins and Nominations}
\usage{
grammys(datamain, a = 52, b = 2010)
}
\arguments{
\item{datamain}{\itemize{
\item The dataset used
}}

\item{a}{The Number of the Grammy Ceremony, For example 52 for 52nd Grammy}

\item{b}{The corresponding year of the ceremony}
}
\value{
Returns a list containing the awards and nominations for each year for upto 10 years - Editable
}
\description{
Grammy Wins and Nominations
}
\examples{
grammys(albumchartssub,52,2010)
}
