% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCol.R
\name{mergecol}
\alias{mergecol}
\title{Merge Responses based on Raw Data}
\usage{
mergecol(a, b)
}
\arguments{
\item{a}{An input dataset}

\item{b}{Response number entered as a string}
}
\value{
The data set with merged responses - UK data (not used in the analysis)
}
\description{
Merge Responses based on Raw Data
}
