% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDescriptives.R
\name{plotDescriptives}
\alias{plotDescriptives}
\title{Creates Descriptive Plots}
\usage{
plotDescriptives(
  workingdir = "~/Dropbox/Projects/MasterThesis",
  results = paste(workingdir, "/results/graphs", sep = "")
)
}
\arguments{
\item{workingdir}{The working directory}

\item{results}{The directory for outputs}
}
\value{
Creates and Saves Descriptive Plots for the Income Distributions of Superstars (2010-2020)
}
\description{
Creates Descriptive Plots
}
