% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotevol.R
\name{plotevol}
\alias{plotevol}
\title{Creates Histogram to Visualize the Evolution of Income}
\usage{
plotevol(workingdir)
}
\arguments{
\item{workingdir}{The working directory}
}
\value{
A Histogram
}
\description{
Creates Histogram to Visualize the Evolution of Income
}
