% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistograms.R
\name{plotHistograms}
\alias{plotHistograms}
\title{Plot Histograms for Indicative Revenue}
\usage{
plotHistograms()
}
\value{
Histograms with Indicative Revenue on the Horizontal Axis (Appendix)
}
\description{
Plot Histograms for Indicative Revenue
}
