% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revHist.R
\name{revHist}
\alias{revHist}
\title{Plot Histograms for Earnings for Music Creators}
\usage{
revHist()
}
\value{
A subset of the main data containing revenue data (UK Data - not used in the analysis)
}
\description{
Plot Histograms for Earnings for Music Creators
}
