% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGrammy.R
\name{scrapegrammy}
\alias{scrapegrammy}
\title{Scrapes Grammy data from Wikipedia}
\usage{
scrapegrammy()
}
\value{
List of all grammy nominations and wins from 2010 - 2020.
}
\description{
Scrapes Grammy data from Wikipedia
}
