% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortResponse.R
\name{sortresponse}
\alias{sortresponse}
\title{Merges all the responses and outputs a dataset}
\usage{
sortresponse()
}
\arguments{
\item{a}{The input dataset}
}
\value{
The data set with merged responses
}
\description{
Merges all the responses and outputs a dataset
}
