% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDescriptives.R
\name{plotDescriptives}
\alias{plotDescriptives}
\title{Plot Descriptive graphs}
\usage{
plotDescriptives()
}
\value{
Creates Descriptive Plots for the Income Distributions of Superstars (2010-2020)
}
\description{
Plot Descriptive graphs
}
