% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistograms.R
\name{plotHistograms}
\alias{plotHistograms}
\title{Plot Histograms}
\usage{
plotHistograms(
  workingdir = "~/Dropbox/Projects/MasterThesis",
  results = paste(workingdir, "/results/graphs", sep = "")
)
}
\arguments{
\item{workingdir}{The working directory}

\item{results}{The directory for outputs}
}
\value{
Histograms with Indicative Revenue on the Horizontal Axis
}
\description{
Plot Histograms
}
